+++ srcjs/custom-controls/layer-switcher.js
@@ -5,10 +5,37 @@ const THEMES = {
   simple: "layer-switcher-ctrl-simple",
 };
 
-function createMenu(layerIds, map) {
+function createLayerLink(map, layerId) {
+  const link = document.createElement("a");
+  link.id = layerId;
+  link.href = "#";
+  link.textContent = layerId;
+  const visibility = map.getLayoutProperty(layerId, "visibility");
+  if (typeof visibility === "undefined" || visibility === "visible") {
+    link.className = "active";
+  }
+
+  link.onclick = function (e) {
+    const layerIdClicked = this.textContent;
+    const visibility = map.getLayoutProperty(layerIdClicked, "visibility");
+    console.log(layerIdClicked, visibility);
+    if (typeof visibility === "undefined" || visibility === "visible") {
+      map.setLayoutProperty(layerIdClicked, "visibility", "none");
+      this.className = "";
+      return;
+    }
+
+    map.setLayoutProperty(layerIdClicked, "visibility", "visible");
+    this.className = "active";
+  };
+  return link;
+}
+
+function createMenu(map, layerIds) {
   const menu = document.createElement("div");
   menu.id = "layer-switcher-menu";
   for (const layerId of layerIds) {
+    /*
     const link = document.createElement("a");
     link.id = layerId;
     link.href = "#";
@@ -31,6 +58,8 @@ function createMenu(layerIds, map) {
       map.setLayoutProperty(layerIdClicked, "visibility", "visible");
       this.className = "active";
     };
+    */
+    const link = createLayerLink(map, layerId);
     menu.appendChild(link);
   }
   return menu;
@@ -48,7 +77,7 @@ export default class LayerSwitcherControl {
     this._container.classList.add(THEMES[this._options.theme || "default"]);
     this._container.style.cssText = this._options.cssText || "";
     const layerIds = this._options.layerIds;
-    this._container.appendChild(createMenu(layerIds, map));
+    this._container.appendChild(createMenu(map, layerIds));
     return this._container;
   }